ai-backend/pyproject.toml
[tool.poetry]
name = "ai-backend"
version = "0.1.0"
description = "AI backend for the Lexon project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11, <3.12"
aiohappyeyeballs = "^2.6.1"
aiohttp = "^3.12.12"
aiosignal = "^1.3.2"
alembic = "^1.16.1"
annotated-types = "^0.7.0"
anyio = "^4.9.0"
appdirs = "^1.4.4"
asgiref = "^3.8.1"
asttokens = "^3.0.0"
attrs = "^25.3.0"
auth0-python = "^4.10.0"
backoff = "^2.2.1"
bcrypt = "^4.3.0"
beautifulsoup4 = "^4.13.4"
blinker = "^1.9.0"
build = "^1.2.2"
cachetools = "^5.5.2"
certifi = "^2025.4.26"
cffi = "^1.17.1"
charset-normalizer = "^3.4.2"
chroma-hnswlib = "^0.7.6"
chromadb = "^0.5.23"
click = "^8.2.1"
cobble = "^0.1.4"
cohere = "^5.15.0"
coloredlogs = "^15.0.1"
crewai = "^0.126.0"
crewai-tools = "^0.47.0"
cryptography = "^45.0.4"
dataclasses-json = "^0.6.7"
decorator = "^5.2.1"
deprecation = "^2.1.0"
distro = "^1.9.0"
docker = "^7.1.0"
docstring-parser = "^0.16"
durationpy = "^0.10"
embedchain = "^0.1.128"
et-xmlfile = "^2.0.0"
executing = "^2.2.0"
fastapi = "^0.115.9"
fastavro = "^1.11.1"
filelock = "^3.18.0"
flatbuffers = "^25.2.10"
frozenlist = "^1.7.0"
fsspec = "^2025.5.1"
google-auth = "^2.40.3"
googleapis-common-protos = "^1.70.0"
gptcache = "^0.1.44"
greenlet = "^3.2.3"
grpcio = "^1.73.0"
h11 = "^0.16.0"
h2 = "^4.2.0"
hf-xet = "^1.1.3"
hpack = "^4.1.0"
httpcore = "^1.0.9"
httptools = "^0.6.4"
httpx = "^0.28.1"
httpx-sse = "^0.4.0"
huggingface-hub = "^0.32.6"
humanfriendly = "^10.0"
hyperframe = "^6.1.0"
idna = "^3.10"
importlib-metadata = "^8.7.0"
importlib-resources = "^6.5.2"
instructor = "^1.8.3"
ipython = "^9.3.0"
ipython-pygments-lexers = "^1.1.1"
jedi = "^0.19.2"
Jinja2 = "^3.1.6"
jiter = "^0.8.2"
json5 = "^0.12.0"
json-repair = "^0.46.2"
jsonpatch = "^1.33"
jsonpickle = "^4.1.1"
jsonpointer = "^3.0.0"
jsonref = "^1.1.0"
jsonschema = "^4.24.0"
jsonschema-specifications = "^2025.4.1"
kubernetes = "^33.1.0"
lancedb = "^0.23.0"
langchain = "^0.3.25"
langchain-cohere = "^0.3.5"
langchain-community = "^0.3.25"
langchain-core = "^0.3.65"
langchain-experimental = "^0.3.4"
langchain-openai = "^0.2.14"
langchain-text-splitters = "^0.3.8"
langsmith = "^0.3.45"
litellm = "^1.68.0"
Mako = "^1.3.10"
mammoth = "^1.9.1"
markdown-it-py = "^3.0.0"
MarkupSafe = "^3.0.2"
marshmallow = "^3.26.1"
matplotlib-inline = "^0.1.7"
mcp = "^1.9.3"
mcp-neo4j-cypher = "^0.2.2"
mcpadapt = "^0.1.9"
mdurl = "^0.1.2"
mem0ai = "^0.1.106"
mmh3 = "^5.1.0"
monotonic = "^1.6"
mpmath = "^1.3.0"
multidict = "^6.4.4"
mypy-extensions = "^1.1.0"
neo4j = "^5.28.1"
networkx = "^3.5"
nodeenv = "^1.9.1"
numpy = "^1.26.4"
oauthlib = "^3.2.2"
onnxruntime = "^1.22.0"
openai = "^1.75.0"
openpyxl = "^3.1.5"
opentelemetry-api = "^1.34.1"
opentelemetry-exporter-otlp-proto-common = "^1.34.1"
opentelemetry-exporter-otlp-proto-grpc = "^1.34.1"
opentelemetry-exporter-otlp-proto-http = "^1.34.1"
opentelemetry-instrumentation = "^0.55b1"
opentelemetry-instrumentation-asgi = "^0.55b1"
opentelemetry-instrumentation-fastapi = "^0.55b1"
opentelemetry-proto = "^1.34.1"
opentelemetry-sdk = "^1.34.1"
opentelemetry-semantic-conventions = "^0.55b1"
opentelemetry-util-http = "^0.55b1"
orjson = "^3.10.18"
overrides = "^7.7.0"
packaging = "^24.2"
pandas = "^2.3.0"
parso = "^0.8.4"
pdfminer-six = "^20250327"
pdfplumber = "^0.11.6"
pexpect = "^4.9.0"
pillow = "^11.2.1"
portalocker = "^2.10.1"
posthog = "^3.25.0"
prompt-toolkit = "^3.0.51"
propcache = "^0.3.2"
protobuf = "^5.29.5"
ptyprocess = "^0.7.0"
pure-eval = "^0.2.3"
pyarrow = "^20.0.0"
pyasn1 = "^0.6.1"
pyasn1-modules = "^0.4.2"
pycparser = "^2.22"
pydantic = "^2.11.5"
pydantic-settings = "^2.9.1"
pydantic-core = "^2.33.2"
Pygments = "^2.19.1"
PyJWT = "^2.10.1"
pypdf = "^5.6.0"
pypdfium2 = "^4.30.1"
PyPika = "^0.48.9"
pyproject-hooks = "^1.2.0"
pyright = "^1.1.402"
pysbd = "^0.3.4"
python-dateutil = "^2.9.0"
python-dotenv = "^1.1.0"
python-multipart = "^0.0.20"
pytube = "^15.0.0"
pytz = "^2025.2"
pyvis = "^0.3.2"
PyYAML = "^6.0.2"
qdrant-client = "^1.14.2"
referencing = "^0.36.2"
regex = "^2024.11.6"
requests = "^2.32.4"
requests-oauthlib = "^2.0.0"
requests-toolbelt = "^1.0.0"
rich = "^13.9.4"
rpds-py = "^0.25.1"
rsa = "^4.9.1"
schema = "^0.7.7"
shellingham = "^1.5.4"
six = "^1.17.0"
sniffio = "^1.3.1"
soupsieve = "^2.7"
SQLAlchemy = "^2.0.41"
sse-starlette = "^2.3.6"
stack-data = "^0.6.3"
starlette = "^0.45.3"
sympy = "^1.14.0"
tabulate = "^0.9.0"
tenacity = "^9.1.2"
tiktoken = "^0.9.0"
tokenizers = "^0.20.3"
tomli = "^2.2.1"
tomli-w = "^1.2.0"
tqdm = "^4.67.1"
traitlets = "^5.14.3"
typer = "^0.16.0"
types-requests = "^2.32.4"
typing-inspect = "^0.9.0"
typing-inspection = "^0.4.1"
typing-extensions = "^4.14.0"
tzdata = "^2025.2"
urllib3 = "^2.4.0"
uv = "^0.7.12"
uvicorn = "^0.34.3"
uvloop = "^0.21.0"
watchfiles = "^1.0.5"
wcwidth = "^0.2.13"
websocket-client = "^1.8.0"
websockets = "^15.0.1"
wrapt = "^1.17.2"
yarl = "^1.20.1"
zipp = "^3.23.0"
zstandard = "^0.23.0"
redis = "^6.2.0"
rq = "^2.4.0"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 